+++ src/System.IO.Compression/tests/System.IO.Compression.Tests.csproj
@@ -10,7 +10,6 @@
   <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'netstandard-Windows_NT-Release|AnyCPU'" />
   <ItemGroup>
     <Compile Include="AsyncStreamTests.cs" />
-    <Compile Include="AsyncStreamTests.netstandard.cs" />
     <Compile Include="DeflateStreamTests.cs" />
     <Compile Include="GZipStreamTests.cs" />
     <Compile Include="XunitAssemblyAttributes.cs" />
